Move DragDrop into Selection
