+++ Apollo/DeviceViewers/GroupViewer.cs
@@ -13,7 +13,6 @@ using Apollo.Binary;
 using Apollo.Components;
 using Apollo.Core;
 using Apollo.Devices;
-using Apollo.DragDrop;
 using Apollo.Elements;
 using Apollo.Helpers;
 using Apollo.Selection;
