+++ Apollo/Selection/Interfaces.cs
@@ -72,6 +72,24 @@ namespace Apollo.Selection {
         int? IExpanded { get; }
         void Expand(int? index);
     }
+    
+    public interface IDroppable: IControl {
+        List<string> DropAreas { get; }
+
+        Dictionary<string, DragDropManager.DropHandler> DropHandlers { get; }
+
+        ISelect Item { get; }
+        ISelectParent ItemParent { get; }
+
+        public bool DropLeft(IControl source, DragEventArgs e)
+            => source.Name == "DropZone" && e.GetPosition(source).Y < source.Bounds.Height / 2;
+    }
+
+    public interface IDraggable: IDroppable, ISelectViewer {
+        string DragFormat { get; }
+
+        void DragFailed(PointerPressedEventArgs e);
+    }
 
     public interface IMutable: ISelect {
         bool Enabled { get; set; }
