+++ Apollo/Viewers/DeviceViewer.cs
@@ -14,7 +14,6 @@ using Avalonia.VisualTree;
 using Apollo.Components;
 using Apollo.Core;
 using Apollo.Devices;
-using Apollo.DragDrop;
 using Apollo.Elements;
 using Apollo.Selection;
