+++ Apollo/Windows/PatternWindow.cs
@@ -16,7 +16,6 @@ using Avalonia.Threading;
 using Apollo.Components;
 using Apollo.Core;
 using Apollo.Devices;
-using Apollo.DragDrop;
 using Apollo.Elements;
 using Apollo.Enums;
 using Apollo.Helpers;
