+++ Apollo/Windows/ProjectWindow.cs
@@ -19,7 +19,6 @@ using Humanizer.Localisation;
 using Apollo.Binary;
 using Apollo.Components;
 using Apollo.Core;
-using Apollo.DragDrop;
 using Apollo.Elements;
 using Apollo.Helpers;
 using Apollo.Selection;
